requirements.txt
Flask
gunicorn==19.9.0
line-bot-sdk==1.14.0
requests
beautifulsoup4
librosa
